requirements.txt
appdirs==1.4.4
astroid==2.12.10
black==21.7b0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.6
coverage==6.5.0
dbfread==2.0.7
dill==0.3.5.1
distlib==0.3.7
filelock==3.12.4
Flask==2.2.2
future==0.18.2
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
importlib-metadata==6.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
mando==0.6.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
pipenv==2023.10.3
platformdirs==3.11.0
protobuf==3.18.1
psycopg2-binary==2.9.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylint==2.15.3
PyNaCl==1.5.0
pyparsing==2.4.7
python-dotenv==1.2.1
radon==5.1.0
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tomli==1.2.1
tomlkit==0.11.5
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.24.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
